# docs/product/v9.2/06-api/openapi.yaml
openapi: 3.1.0
info:
  title: Xingjing Drama Studio API
  version: 9.2.0-spec.1
  description: 星镜剧创 v9.2 四端 Web 开发契约。所有租户接口均需服务端校验工作区和对象归属。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
...
paths:
...
  /projects/{projectId}/exports:
    get:
      tags: [Export]
      summary: list Project Exports
      operationId: listProjectExports
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '400': {$ref: '#/components/responses/ValidationError'}
        '200': {$ref: '#/components/responses/ListResponse'}
    post:
      tags: [Export]
      summary: create Export Task
      operationId: createExportTask
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody: {$ref: '#/components/requestBodies/EntityRequest'}
      responses:
        '202': {$ref: '#/components/responses/ObjectResponse'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /projects/{projectId}/compliance:
    get:
      tags: [Compliance]
      summary: get Project Compliance
      operationId: getProjectCompliance
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '400': {$ref: '#/components/responses/ValidationError'}
        '200': {$ref: '#/components/responses/ObjectResponse'}
  /projects/{projectId}/compliance/actions:
    post:
      tags: [Compliance]
      summary: execute Compliance Action
      operationId: executeComplianceAction
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody: {$ref: '#/components/requestBodies/ActionRequest'}
      responses:
        '202': {$ref: '#/components/responses/ObjectResponse'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminSession:
      type: apiKey
      in: cookie
      name: xj_admin_session
    reviewToken:
      type: apiKey
      in: cookie
      name: xj_review_session
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
...
  requestBodies:
    ActionRequest:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ActionRequest'}
    EntityRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  responses:
    ObjectResponse:
      description: 单对象成功响应
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ObjectResponse'}
    ListResponse:
      description: 游标分页列表响应
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ListResponse'}
    ValidationError:
      description: 请求字段不合法
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    Conflict:
      description: 版本、状态、计费、幂等或合规冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    Uuid:
      type: string
      format: uuid
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId: {$ref: '#/components/schemas/Uuid'}
        serverTime: {type: string, format: date-time}
        page:
          type: object
          properties:
            size: {type: integer}
            nextToken: {type: [string, 'null']}
    ObjectResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: object, additionalProperties: true}
        meta: {$ref: '#/components/schemas/Meta'}
    ListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {type: object, additionalProperties: true}
        meta: {$ref: '#/components/schemas/Meta'}
    ErrorDetail:
      type: object
      properties:
        field: {type: string}
        reason: {type: string}
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            details:
              type: array
              items: {$ref: '#/components/schemas/ErrorDetail'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    ActionRequest:
      type: object
      required: [action]
      properties:
        action: {type: string, minLength: 1, maxLength: 80}
        targetId: {$ref: '#/components/schemas/Uuid'}
        payload: {type: object, additionalProperties: true}
...